% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacroix_palettes.R
\docType{data}
\name{lacroix_palettes}
\alias{lacroix_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 21.}
\usage{
lacroix_palettes
}
\description{
Complete list of palettes
}
\keyword{datasets}
