% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_lacroix}
\alias{scale_color_lacroix}
\alias{scale_colour_lacroix}
\alias{scale_fill_lacroix}
\title{LaCroix color scales}
\usage{
scale_color_lacroix(name, discrete = FALSE, ...)

scale_colour_lacroix(name, discrete = FALSE, ...)

scale_fill_lacroix(name, discrete = FALSE, ...)
}
\arguments{
\item{name}{Name of color palette (LaCroix flavor) desired.}

\item{discrete}{Generate a discrete palette? Defaults to \code{FALSE}, aka continuous mapping.}

\item{...}{parameters passed on to \link[ggplot2]{\code{discrete_scale}} (discrete mapping)
or \link[ggplot2]{\code{scale_color_gradientn}} (continuous mapping).}
}
\value{
A Scale object.
}
\description{
Create color scales using LaCroix palettes.
Currently only works with the flavor-specific palettes (not the "paired" palette).
}
\details{
Implementation heavily inspired by the \link[viridis] package.
}
\examples{
library(ggplot2)

# ggplot2 examples, but with more seltzer
p <- ggplot(mtcars, aes(wt, mpg))
p + geom_point(size=4, aes(color = factor(cyl))) +
    scale_color_lacroix("PassionFruit", discrete = TRUE) +
    theme_bw()

p + geom_point(size=4, aes(color = hp)) +
    scale_color_lacroix("PassionFruit", discrete = FALSE) +
    theme_bw()

}
\keyword{color}
\keyword{palettes}
\keyword{scales}
