% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaCroixColor.R
\name{lacroix_palette}
\alias{lacroix_palette}
\title{A LaCroix color palette generator}
\usage{
lacroix_palette(name, n, type = c("discrete", "continuous", "paired"))
}
\arguments{
\item{n:}{Number of colors desired If omitted, uses all colours.}

\item{name:}{Number of colors desired If omitted, uses all colours.}

\item{type:}{Either "discrete", "continuous", or "paired".}
}
\description{
This function allows you to generate color palettes based on LaCroix water flavors.
}
\examples{
lacroix_palette()
}
\keyword{color}
\keyword{palettes}
