% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.palette.R
\name{print.palette}
\alias{print.palette}
\title{Function to print color palettes}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{n:}{Number of colors desired If omitted, uses all colours.}

\item{name:}{Number of colors desired If omitted, uses all colours.}
}
\description{
Function to print color palettes
}
